LayoutTests/imported/w3c/web-platform-tests/css/css-values/attr-revert-rule-expected.txt


FAIL The revert-rule works in an attr() fallback assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
